+++ feature/received/src/main/java/com/susu/feature/received/navigation/ReceivedNavigation.kt
@@ -20,7 +20,7 @@ import com.susu.feature.received.ledgeredit.LedgerEditRoute
 import com.susu.feature.received.ledgerfilter.LedgerFilterRoute
 import com.susu.feature.received.navigation.argument.FilterArgument
 import com.susu.feature.received.received.ReceivedRoute
-import com.susu.feature.received.search.LedgerSearchRoute
+import com.susu.feature.received.ledgersearch.LedgerSearchRoute
 import kotlinx.serialization.json.Json
 
 fun NavController.navigateReceived(navOptions: NavOptions) {
